I modified testtools/td/scripts/Test so that it works in the directory tree as released. batchvision passes all pieces of the testdeck, except the final component  truncate.S.
auditshell.c was an older version. I updated that so that release information could be embedded in the executables.

Removed the g0 from CC command line. With it, the batchvision executable was 300MB (vs our 7.1.7 release of 18MB)

Dropped the optimization level from 5 to 3 (the default) so that 123.45 printWithCommas: -10.5 would work without a Read-Eval-Print error.

Removed the 32bit version from the default targets for the debug mode build (it didnt work, and we dont use 32bit anymore, anyway)