Restore logic to limit parallel make resource use.
Commit fec7f23b45e6a46d8076d8bc0e90b8a42309a35c removed too much of the logic
that limited the resources consumed by parallel makes (GNU make -j).  While
not an issue for systems with lots of CPUs and memory (especially memory),
it is a problem for smaller systems.  This commit strikes a compromise.  Now,
if the CPUCOUNT environment isn't set, the number of CPUs used for parallel
makes will be the number of CPUs in the system.  If the CPUCOUNT environment
variable is set and numeric, parallel makes will be limited to use no more
than that number of CPUS.  If CPUCOUNT is set and non-numeric (e.g. export
CPUCOUNT=unlimited), CPU use will be unrestricted.