Reinitialize MAKEOPTS before recomputing them.
'visionBuilder' calls 'getMakeOpts' multiple times per build.  This is
unnecessary but probably easier to leave as is for now.  At least make
sure that it clears the previous value of the MAKEOPTS shell variable
before updating it so that the options string doesn't keep accumulating
multiple copies of the same options.