Fix at-exit crashes of 8.1 executables on linux.
The V library is responsible for memory management for all Vision
executables.  When a library is unloaded at exit as opposed to being
simply abandoned (as it is on linux), the static C++ objects it
contains are destroyed.  In the case of V, if those static objects
require the services of the memory management facilities of V, a
destruction order matters.  In particular, if the static objects
used by the memory manager are destroyed before static objects that
will use the memory manager, bad things happen.  That was the case
in the 8.1 version of V, which placed the definition of a static
VString member of class VInfoServer ahead of the definitions for the
memory manager's static variables in the file.  Since static destructors
are run in reverse order of their appearance in a file to insure their
destruction in the opposite order of their construction, the all
executables that loaded V crashed at exit.  Two fixes were possible -
move the definition or change the data type of the offending member
to eliminate the need for offending destructor call.  This fix does
takes the latter approach by changing an array of VString objects
into an array of pointers to const chars.