Remove Solaris symbol equivalences not needed in x86 builds.
A bug in some version of the Solaris c++ compiler occasionally caused the
compiler to generate different name manglings for same object, resulting
in undefined symbol linker errors for some references to those objects.
The workaround for that problem is to explicitly equivalence the mangled
names using #pragma weak directives (see the code for a full explanation).
Some of those equivalences are no longer needed for x86 builds using
Solaris Studio 2.1 (they remain for the benefit of sparc builds which
still appear to need them).