VString 'int'->'size_t' and 'clang' related cleanups.
While addressing the 'clang' complaint described below, several
variables that were defined to be 'int's were when they should have been
'size_t's were found and corrected.

'clang' warns when the test in an 'if' or 'while' statement uses an
assignment (i.e., 'var = expr') to simultaneous capture and test the
value of an expression.  This commit addresses that complaint.