Shared libraries also WORK differently on the Mac.
On the Mac, shared libraries need to have an installation path
baked into them when they are linked.  That path can be relative
to the executable or library that's loading them but it must
include the name of the library, not just the path to the
directory that contains it.  For example, the command that builds
a 'libV' that can be found by other executables and libraries in
our directory tree needs to include the option:

	-install_name @loader_path/../lib/libV.dylib

Finding a way to generate that option in a platform-independent
makefile turns out to be a little harder than it seems, so for
now, this commit hacks that capability into 'VLINK-CC'.