The first working version of frontend window resizing.
It's alive.  A bit more work is needed to figure out why the screen repaints
correctly most, but not all, of the time, especially when it's being made
smaller.

The basic approach to making this work involves processing KEY_RESIZE events.
Note that this only works with 'ncurses' based libraries (the Mac and Linux)
and will not work out of the box on Solaris, which requires that we provide
a handler for the SIGWINCH signal.

The heavy lifting in getting this to work at all was tracking down which of
the many global variables used by the editor needed to be eliminated or
recomputed.