Add (but disable) terminal resize handling on Solaris.
The default version of 'curses' on Solaris is old.  In particular, it is not
'ncurses', which provides support for terminal window resizing.  In theory,
it's possible to support resizing by catching the SIGWINCH but given the
bookkeeping involved and the fragility of the frontend code base, that's
easier said than done.  Since resizing now works with only minor glitches on
Linux and the Mac, it's just not worth the effort to go down the Solaris
path any further.