Make a Vxa adapter (pugixml) work on Windows
This commit implements our first VXA adapter on Windows.  That is
significant not because of what the adapter does but rather because
of the changes to Vxa needed to make it work in an environment where
the compiler and linker hide shared library symbols by default.  This
isn't a new issue.  This code base has long been structured to support
the hidden symbol default of Windows and the visible symbol default of
Linux. Until now, we simply never debugged the Vxa structures used to
support adapter development in a default hidden symbol world.  This
commit addresses that.

Even though symbol visibility defaults are different on Linux and Windows,
it is possible to compile and link code on Linux that hides symbols by
default.  At least one Linux build of this code does just that.  In all
likelihood, that build encountered exactly the same problems as those
addressed here.  For a variety of reasons including incompatibilities
between the Linux motivated fixes and the changes required to make this
code work on Windows, those changes have been reverted (see this commit
and commit 2d72345).  If, as expected, the problems resolved here turn
out to be the same as those resolved in the earlier Linux build, this
version of the code should (or can be made to) address the Linux build's
problems as well.

In a nutshell, hiding symbols by default has the potential to subvert
enforcement of C++'s one-definition rule.  Specifically, C++'s one-
definition rule requires that there be exactly one instance of a symbol's
definition at runtime.  That requirement is GLOBAL, spanning not only the
executable being run, but all shared libraries to which it is linked.
With the compiler issuing guidance, enforcement responsibility falls
mainly to the linker.  When symbols are hidden, the linker may lack the
information it needs to properly do its job.

While normally not an issue for function definitions, it can be an issue
for static member variables of templated classes.  Such clases are used
extensively by Vxa (and Vca) to make type information introspectable at
run time (e.g., Vxa::VExportable<T>::g_pTraits).  The associated issues
arise because of how the compiler generates and annotates template
instantiations.  Unlike non-templated objects, the compiler typically
generates multiple definitions of templated objects, one for each implicit
instantiation of the template.  Because multiple instantiations of the
same objects are generated, each generated definition is tagged as a weak
definition of the associated object.  In contrast, definitions of non-
templated objects (and explicitly instantiated templates) are emitted
exactly once and tagged as strong definitions.  The linker ensures that
there is no more than one strong definition of an object and that weak
and strong definitions of the same object are merged to produce a single
definition in the resultant library or executable (those definitions are
marked as weak if all source definitions are weak, strong if a strong
definition is present).  When all symbols are visible by default, this
approach properly enforces the one-definition rule.  When symbols are
hidden by default, the code must be properly structured and tagged to
ensure that a single strong definition is explicitly created and marked
as visible.

In the case of Vxa, two sets of template instantiations are improperly
exported.  Those instantiations provide type information for standard
numeric and string data types.  Previously found in Vxa_VImportable.cpp
and Vxa_VExportable.cpp, this commit moves them to visible locations
in Vxa_VImportable.h and Vxa_VExportable.h.  In the process, it also
gives them correct export tagging.

Prior to this move, these instantiations were properly initialized but
were strictly private to the Vxa library.  As a consequence, adapters
that relied on these instantiations (e.g., Vxa::VExportable<VString>,
Vxa::VImportable<VString>, etc.) did not see and link to the official
instantiations in Vxa but instead created new, distinct, UNINITIALIZED,
instantiations of their own.  Because those new instantiations were not
initialized, they simply did not work!

By way of anatomy, here's an example of one instantiation declaration
from Vxa_VExportable.h:

  Vxa_VExportable_Implementation template class Vxa_API Vxa::VExportable<int>;

Both Vxa_VExportable_Implementation and Vxa_API are macros that are
expanded differently depending on where the declaration is encountered.
In Vxa_VExportable.cpp, this declaration is used to actually instantiate
the template.  There, this declaration expands to:

  template class __declspec(dllexport) Vxa::VExportable<int>;

resulting in both a template instantiation and directives to the linker
to export all objects created as part of that instantiation.

When Vxa_VExportable.h is included elsewhere, however, this declaration
expands to either:

  extern template class __declspec(dllexport) Vxa::VExportable<int>;

when included while compiling other components used to build the Vxa
library or:

  extern template class __declspec(dllimport) Vxa::VExportable<int>;

everywhere else.  When compiling Vxa, the result is the creation of
exactly one, fully exported instantation of Vxa::VExportable<int>.  When
encountered anywhere else, the result is creation of external references
to these symbols associated with the instantiation that will ultimately
be resolved by the linker to the official definitions found in Vxa.

The changes outlined above are the essential changes.  Two other sets of
changes round out this commit.  Some inlined members of VExportable<T>
make use of another templated Vxa class - VMonotypeMapMakerFor<T>.  The
instantiations of that template for the same set of built-in types needed
to be exported from Vxa.  This commit does that.

Last but not least, there was one 'misuse' of Vxa_API.  The template
definitions of VExportable and VImportable were tagged as being part
of Vxa (e.g., template <typename T> class Vxa_API VExportable {...}).
Additionally, the macros used to initiate VExportable and VImportable
for new types were also marked as being part of Vxa's interface.  As
a result, at least on Windows, the compiler blocked instantiations of
those templates everywhere but inside Vxa itself - not the intended
behavior.  Corrected here as well.