Eliminate 'this' pointer null test in 'VSymbolBinding::level ()'
GCC 6.3 (at least) no longer generates code that compares 'this' pointers to
'null'.  While it's trivially easy to invoke a non-virtual member through using
a 'null' pointer, the standard considers this to be 'undefined' behavior, so
GCC 6.3 silently refuses to generate code for the test.  While there may be
clever ways to circumvent the compiler on the matter, the safest approach is to
reorganize the code to give the compiler what it wants to see.  This commit does
that by relocated the required tests in the case of the implementation of
'VSymbolBinding::level ()' in the backend.