Address newly absent 'matherr' structures on Linux
As of 'glibc' 2.27 (see issue #48), Linux no longer supports 'matherr'.
For our purposes, the implication of that change is the disappearance
of a definition for 'struct __exception' from the <math.h> header files.
For portability reasons, our code long ago began using 'isfinite' to
detect the issues 'matherr' was designed to address.  For legacy reasons,
we continue to provide an implementation of a 'matherr' callback to ensure
that very old and System V Unix libraries return results detectable by
'isfinite'.  Our Linux code is known to work correctly with and without
it.  That being the case, we just need to make sure that the virtually
unused callback we provide continues to compile and could work if it was
ever called.  We've already addressed that case for other platforms.  Add
Linux to the list.