Remove absolute link to GitHub Pages website
This file is an input to a Github Pages build process that can be used in any fork or upstream of this repository.  Even though this file contains links that currently don't work when viewed as local markdown (see Jekyll docs for possible workarounds), it makes little sense to include an absolute link to one particular Github Pages website that will be published in every Github Pages site built from this file.